requirements.txt
django==4.1.3
dj-rest-auth==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
dj-rest-auth[with_social]
djangorestframework-simplejwt==5.2.2
django-crispy-forms==1.14.0










